Purge command

+++ cogs/general.py
@@ -8,15 +8,25 @@ class GeneralCommands(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    @commands.command(name='ping', aliases=['poing'])
+    @commands.command(name='ping',
+                      aliases=['poing']
+                      )
     async def ping(self, ctx):
         await ctx.send(f"{ctx.message.author.mention} **PONG!**\nI took `{round(self.bot.latency * 1000)} ms` to respond")
 
     @commands.command(name='feedback', aliases=['complain', 'compliment', 'suggest'], hidden=True)
     async def feedback(self, ctx, *args):
-        feedbackChannel = self.bot.get_channel(779738194097995806)
-        await feedbackChannel.send(f"{ctx.author.id} (**{ctx.message.author.name}#{ctx.message.author.discriminator}**) says ```css\n{' '.join(args)}```")
-        await ctx.message.delete()
+        if (not await self.bot.is_owner(ctx.message.author)):
+            feedbackChannel = self.bot.get_channel(779738194097995806)
+            await feedbackChannel.send(f"{ctx.author.id} (**{ctx.message.author.name}#{ctx.message.author.discriminator}**) says ```css\n{' '.join(args)}```")
+            await ctx.message.delete()
+
+    @commands.command(name='purge',
+                      decription="Bulk delete messages"
+                      )
+    @commands.has_permissions(manage_messages=True)
+    async def mass_delete(ctx, self, amount):
+        await ctx.channel.purge(limit=amount)
 
 
 def setup(bot):
